# dify/openapi_tools.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Deep Research Agent — Memory Service",
    "version": "1.0.0",
    "description": "Memory management endpoints for the research agent workflow"
  },
  "servers": [
    {
      "url": "http://host.docker.internal:8100"
    }
  ],
  "paths": {
    "/memory/store": {
      "post": {
        "operationId": "storeMemory",
        "summary": "Store a summarised research chunk and update the token budget",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["text", "query"],
                "properties": {
                  "text": { "type": "string", "description": "The text to store" },
                  "query": { "type": "string", "description": "The sub-question this text answers" },
                  "session_id": { "type": "string", "description": "Session identifier" },
                  "metadata": { "type": "object", "description": "Optional metadata" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Stored successfully" }
        }
      }
    },
    "/memory/retrieve": {
      "post": {
        "operationId": "retrieveMemory",
        "summary": "Retrieve the most relevant stored summaries for a query",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["query"],
                "properties": {
                  "query": { "type": "string", "description": "The query to search for" },
                  "top_k": { "type": "integer", "default": 5, "description": "Number of results" },
                  "session_id": { "type": "string", "description": "Filter by session" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Retrieved results" }
        }
      }
    },
    "/memory/budget": {
      "get": {
        "operationId": "getBudget",
        "summary": "Return current token usage vs budget",
        "responses": {
          "200": { "description": "Budget status" }
        }
      }
    },
    "/memory/reset": {
      "post": {
        "operationId": "resetSession",
        "summary": "Reset the session token budget",
        "responses": {
          "200": { "description": "Budget reset" }
        }
      }
    },
    "/memory/summarize": {
      "post": {
        "operationId": "summarizeText",
        "summary": "Compress text within the token budget using the LLM",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["text"],
                "properties": {
                  "text": { "type": "string", "description": "Text to summarise" },
                  "max_tokens": { "type": "integer", "description": "Max tokens for summary" },
                  "focus": { "type": "string", "description": "Focus hint for summarisation" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Summary result" }
        }
      }
    },
    "/decompose": {
      "post": {
        "operationId": "decomposeQuery",
        "summary": "Break a complex research question into sub-questions",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["query"],
                "properties": {
                  "query": { "type": "string", "description": "Complex query to decompose" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Decomposed sub-questions" }
        }
      }
    },
    "/research": {
      "post": {
        "operationId": "researchSubquery",
        "summary": "Execute a single sub-query: research, summarise, and store",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["subquery"],
                "properties": {
                  "subquery": { "type": "string", "description": "The sub-question to research" },
                  "session_id": { "type": "string", "description": "Session identifier" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Research result" }
        }
      }
    },
    "/synthesize": {
      "post": {
        "operationId": "synthesize",
        "summary": "Retrieve context and synthesise a final answer",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["query"],
                "properties": {
                  "query": { "type": "string", "description": "Original user query" },
                  "top_k": { "type": "integer", "default": 5 },
                  "session_id": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Final synthesised answer" }
        }
      }
    }
  }
}